+++ src/main/java/kitchenpos/product/dto/ProductRequest.java
@@ -1,8 +1,8 @@
-package kitchenpos.menu.dto;
+package kitchenpos.product.dto;
 
 import java.math.BigDecimal;
 
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public class ProductRequest {
