+++ src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.repository;
+package kitchenpos.product.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kitchenpos.menu.domain.Product;
+import kitchenpos.product.domain.Product;
 
 public interface ProductRepository extends JpaRepository<Product, Long> {}
